+++ changelog.md
@@ -3,6 +3,7 @@
 __                                                                                                      __
 ### Mains
 * Airdrops And Parachutes
+* Downeds Animation
 __                                                                                                      __
 ### Menu
 __                                                                                                      __